heyken/heyken_bot/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "heyken_bot"
version = "0.1.0"
description = "Bot integrujący RocketChat z Ollama do realizacji projektów"
readme = "README.md"
authors = [
    {name = "Heyken Team", email = "info@heyken.local"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "pydantic>=2.0.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "docker>=6.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
